* Fix bugs found in a full audit of src, inc and drv
Compile errors:
- statistic.h: missing semicolon on the standardDeviationi32 prototype
- circBuf.c: NULL used without stddef.h
- comat.c, comstxetx.c: the driver variable was spelled drive
- comstxetx.c: bound checked a rxMaxLength member that does not exist
- basicmatrix.c: undeclared jSize, and limitUpDw2D did not match its prototype

Silent logic bugs:
- search.c: a stray semicolon made linearSearch always report index 0
- comat.c, comstxetx.c: Evaluate assigned instead of compared the ready flag
- basicmath.c: findMini32 returned the maximum element
- bininp.c: the rising field was never set, so the getter always read FALSE
- pid.c: pidInit left the error memory and the term accumulators uninitialized
- complex.c: complexDiv divided only the second term; polar used atan
- basicmatrix.c: the 2D row stride was the row count instead of the column count
- basicmath.c: calculateMedian did not average an even sized array

Bounds and divide by zero:
- sort.c, search.c: length - 1 underflowed for a zero length array
- basicmath.c, statistic.c: zero length and divide by length guards added

Header collisions and API mismatches:
- comsec.h: the typedef reused the comat_t name
- dcmotor.h, complex.h, logic.h: duplicate include guards
- maf.c: the u32 variants declared by the header were missing
- comat.h: packetProcess took rxInd by pointer while callers passed a value
- dcmotor: the pwm callback takes float, matching the rest of the library

Also renamed comssafe.c to comsafe.c, fixed the dcmotor banner, made
hc597DrvOneShoot honour driver->size, and updated PID_Test to the current
pidInit signature.

All sources compile clean under -Wall -Wextra except the four known empty
driver stubs. All headers coexist in one translation unit. All tests link.

+++ drv/dcmotor.c
@@ -1,16 +1,21 @@
 /**
   ******************************************************************************
   *
-  * @file:      hc597.c
+  * @file:      dcmotor.c
   * @author:    Engin Subasi
   * @email:     enginsubasi@gmail.com
   * @address:   github.com/enginsubasi
   *
-  * @version:   v 0.0.1
+  * @version:   v 0.0.2
   * @cdate:     23/05/2022
   * @history:   23/05/2022 Created
+  *             29/07/2026 The file banner named the wrong file and device. It
+  *                        was copied from the hc597 driver.
+  *             29/07/2026 The pwm callback takes float instead of double, to
+  *                        match the rest of the library. This changes the public
+  *                        API.
   *
-  * @about:     HC597 driver file.
+  * @about:     DC motor driver file.
   * @device:    Generic
   *
   * @content:
@@ -31,13 +36,13 @@
 void dcMotorInit ( struct dcmotor_t *driver,
                     void ( *bridgeHighFnc )( uint8_t ),
                     void ( *bridgeLowFnc )( uint8_t ),
-                    void ( *pwmFnc )( double ))
+                    void ( *pwmFnc )( float ))
 {
     driver->bridgeHigh = bridgeHighFnc;
     driver->bridgeLow = bridgeLowFnc;
     driver->pwm = pwmFnc;
 
-    driver->pwm ( ( double ) 0.0 );
+    driver->pwm ( ( float ) 0.0 );
     dcMotorBridgeState ( driver, BRIDGE_NO );
 }
